Various string and array optimizations.
[gtk] RomDataView: Use a string table for date formats.

[gtk] ConfigDialog: Change the icon size array from `int` to `uint8_t`.
This code path is not used on GTK4, so we're not including the code size
difference for GTK4 ConfigDialog.

[librpbase] TextOut_text.cpp: Use a string table for date formats.

[libromdata]
- NES: Use a fixed-length character array for footer mappers.
- Nintendo3DSFirm: Use a fixed-length character array for sighax status.

Code size differences: (64-bit Gentoo Linux, gcc-12.1.0, release build, no LTO)

   text    data     bss     dec     hex filename
  80176      72      56   80304   139b0 RomDataView.cpp.o [XFCE] [before]
  80219       8      56   80283   1399b RomDataView.cpp.o [XFCE] [after]
    +43     -64       0     -21     -15 Difference

   text    data     bss     dec     hex filename
  81329      72      88   81489   13e51 RomDataView.cpp.o [GTK3] [before]
  81388       8      88   81484   13e4c RomDataView.cpp.o [GTK3] [after]
    +59     -64       0      -5      -5 Difference

   text    data     bss     dec     hex filename
  80132      72      88   80292   139a4 RomDataView.cpp.o [GTK4] [before]
  80191       8      88   80287   1399f RomDataView.cpp.o [GTK4] [after]
    +59     -64       0      -5      -5 Difference

   text    data     bss     dec     hex filename
  12096       8      24   12128    2f60 ConfigDialog.cpp.o [XFCE] [before]
  12004       8      24   12036    2f04 ConfigDialog.cpp.o [XFCE] [after]
    -92       0       0     -92     -5c Difference

   text    data     bss     dec     hex filename
  10879       8      24   10911    2a9f ConfigDialog.cpp.o [GTK3] [before]
  10811       8      24   10843    2a5b ConfigDialog.cpp.o [GTK3] [after]
    -68       0       0     -68     -44 Difference

   text    data     bss     dec     hex filename
  31769      72       1   31842    7c62 TextOut_text.cpp.o [before]
  31998       8       1   32007    7d07 TextOut_text.cpp.o [after]
   +229     -64       0    +165     +a5 Difference

   text    data     bss     dec     hex filename
  20646    1176       0   21822    553e NES.cpp.o [before]
  20678    1136       0   21814    5536 NES.cpp.o [after]
    +32     -40       0      -8      -8 Difference

   text    data     bss     dec     hex filename
   8746     704       0    9450    24ea Nintendo3DSFirm.cpp.o [before]
   8802     584       0    9386    24aa Nintendo3DSFirm.cpp.o [after]
    +56    -120       0     -64     -40 Difference

I'm not entirely sure why TextOut_text.cpp.o grew so much compared to
GTK RomDataView.cpp.o...

+++ src/gtk/config/ConfigDialog.cpp
@@ -114,7 +114,7 @@ config_dialog_init(ConfigDialog *dialog)
 
 	// Set the window icon.
 	// TODO: Redo icon if the icon theme changes?
-	const int icon_sizes[] = {16, 32, 48, 64, 128};
+	const uint8_t icon_sizes[] = {16, 32, 48, 64, 128};
 	GList *icon_list = nullptr;
 	for (int icon_size : icon_sizes) {
 		GdkPixbuf *const icon = gtk_icon_theme_load_icon(
